+++ lib/cequel/record.rb
@@ -1,5 +1,6 @@
 # -*- encoding : utf-8 -*-
 require 'active_model'
+require 'weakref'
 
 require 'cequel'
 require 'cequel/record/errors'
@@ -116,6 +117,29 @@ module Cequel
       #
       def establish_connection(configuration)
         self.connection = Cequel.connect(configuration)
+      end
+
+      # @return [Array<Class>] All the record classes that are
+      #   currently defined.
+      def descendants
+        weak_descendants.map do |clazz|
+          begin
+            clazz.__getobj__ if clazz.weakref_alive?
+          rescue WeakRef::RefError
+            nil
+          end
+        end.compact
+      end
+
+      # Hook called when new record classes are created.
+      def included(base)
+        weak_descendants << WeakRef.new(base)
+      end
+
+      private
+
+      def weak_descendants
+        @weak_descendants ||= []
       end
     end
   end