+++ lib/cequel/spec_support.rb
@@ -1,0 +1,11 @@
+module Cequel
+  #
+  # A collection of modules providing convenient functionality for running
+  # test suites with Cequel
+  #
+  # @see Preparation
+  #
+  module SpecSupport
+    autoload :Preparation, "cequel/spec_support/preparation"
+  end
+end