+++ spec/examples/record/list_spec.rb
@@ -9,7 +9,7 @@ describe Cequel::Record::List do
     list :contributor_ids, :int
   end
 
-  let(:scope) { cequel[:posts].where(:permalink => 'cequel') }
+  let(:scope) { cequel[Post.table_name].where(:permalink => 'cequel') }
   subject { scope.first }
 
   let! :post do